+++ source/components/table/table.js
@@ -22,6 +22,7 @@ var TableDefaultConfig = {
 
     head: null,
     body: null,
+    static: false,
     source: null,
 
     searchMinLength: 1,
@@ -210,6 +211,22 @@ var Table = {
             this.searchFields = Utils.strToArray(o.searchFields);
         }
 
+        if (Utils.isValue(o.head)) {
+            o.head = Utils.isObject(o.head);
+        }
+
+        if (Utils.isValue(o.body)) {
+            o.body = Utils.isObject(o.body);
+        }
+
+        if (o.static === true) {
+            o.showPagination = false;
+            o.showRowsSteps = false;
+            o.showSearch = false;
+            o.showTableInfo = false;
+            o.rows = -1;
+        }
+
         if (o.source !== null) {
             Utils.exec(o.onDataLoad, [o.source], element[0]);
             element.fire("dataload", {
@@ -256,6 +273,14 @@ var Table = {
         this.items = [];
         this.heads = [];
         this.foots = [];
+
+        if (Array.isArray(o.head)) {
+            this.heads = o.head;
+        }
+
+        if (Array.isArray(o.body)) {
+            this.items = o.body;
+        }
 
         if (Utils.isValue(data)) {
             this._createItemsFromJSON(data);
@@ -1681,7 +1706,7 @@ var Table = {
     setHeadItem: function(name, data){
         var i, index;
         for(i = 0; i < this.heads.length; i++) {
-            if (item.name === name) {
+            if (this.heads[i].name === name) {
                 index = i;
                 break;
             }
@@ -1696,9 +1721,19 @@ var Table = {
     },
 
     setData: function(/*obj*/ data){
+        var o = this.options;
+
         this.items = [];
         this.heads = [];
         this.foots = [];
+
+        if (Array.isArray(o.head)) {
+            this.heads = o.head;
+        }
+
+        if (Array.isArray(o.body)) {
+            this.items = o.body;
+        }
 
         this._createItemsFromJSON(data);
 
@@ -1738,6 +1773,15 @@ var Table = {
                     source: o.source,
                     data: data
                 });
+
+                if (Array.isArray(o.head)) {
+                    that.heads = o.head;
+                }
+
+                if (Array.isArray(o.body)) {
+                    that.items = o.body;
+                }
+
                 that._createItemsFromJSON(data);
                 that._rebuild(review);
             }, function(xhr){