+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 + [x] M4Q Ajax: fix creating additional headers for request, issue #1427
 + [x] Init: add meta tag `metro4:init:mode`. The value must be `immediate` or `contentloaded`. If value is `contentloaded` and meta tag `metro4:init` not equal  `false`, Metro 4 will be initiate after the content was loaded.
 + [x] Table: now you can define `decimalSeparator` and `thousandSeparator` in table head and global
++ [x] Table: add attributes `data-head` (object name, define table header), `data-body` (object name, define table data)
++ [x] Table: add attribute `data-static`. If `true`, search, pagination, info, rows count will be disabled and will not showing.
 
 ### 4.3.1
 + [x] Interop: fix interop with Metro4 and jQuery, issue #1408